requirements.txt
backports.ssl-match-hostname==3.5.0.1
six==1.10.0
websocket-client==0.35.0
wheel==0.24.0
